requirements.txt
pywebview[cocoa]
flask
flask-wtf